+++ systeme/server/game_test/MapHandler.cpp
@@ -709,13 +709,13 @@ void MapHandler::playerInit(PlayerShip* p1,PlayerShip* p2) {
 }
 void MapHandler::updateBounds() {
     for( PlayerShip* p : player_ships_set){
-        p->setBounds({ { static_cast<uint_fast16_t>(p->getPos().x -1), p->getPos().y}, {3, 2}});
+        p->setBounds({ { static_cast<int>(p->getPos().x -1), p->getPos().y}, {3, 2}});
     }
     for(EnemyShip* e: enemy_ships_set){
-        e->setBounds({{static_cast<uint_fast16_t>(e->getPos().x-1), e->getPos().y},{3,1}});
+        e->setBounds({{static_cast<int>(e->getPos().x-1), e->getPos().y},{3,1}});
     }
     for(Boss* b: boss_set){
-        b->setBounds({{static_cast<uint_fast16_t>(b->getPos().x), b->getPos().y},{18,6}});
+        b->setBounds({{static_cast<int>(b->getPos().x), b->getPos().y},{18,6}});
     }
 
 }
